Add support for building Wasm with emscripten
Building the C++ bindings does not work because of what seems to be a bug in
emscripten where it fails to parse its own metadata.  I have not tracked this
down yet, it may even be fixed already in a later version than the one in the
Docker image.  Everything required to (almost) build the C++ bindings is
included in this commit regardless.