Remove ExessBlob
In retrospect, the added indirection of a struct that contains a pointer makes
things confusing and inconsistent.  This way, bounded and unbounded types may
both be written to buffers directly in roughly the same way.  The cost is that
the function signatures for unbounded types are slightly different because of
the additional size parameter.