Fix main man page formatting
These constructs produced output that was incorrectly parsed (by either `man`
or `Man` mode in emacs). I'm not sure why, but work around these issues by
sprinkling in some newlines and changing "Binary" to "Binary Blobs" which for
whatever reason is correctly interpreted as a definition list entry.

With this, all of the cross-references in the "SEE ALSO" section are correctly
displayed by the standard tools, and function properly as hyperlinks in
emacs (and presumably any other reader that supports this).